pyproject.toml
[tool.poetry]
name = "recoani-api"
version = "0.1.0"
description = ""
authors = ["fastsnowy", "re-nan6"]
license = "MIT"
readme = "README.md"
packages = [{include = "recoani_api"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.89.1"
fastapi-csrf-protect = "^0.2.2"
gql = "^3.4.0"
graphql-core = "^3.2.3"
motor = "^3.1.1"
numpy = "^1.24.1"
pandas = "^1.5.2"
pymongo = "^4.3.3"
python-decouple = "^3.7"
scikit-learn = "^1.2.0"
uvicorn = "^0.20.0"
gunicorn = "^20.1.0"
implicit = "^0.6.2"
requests = "^2.28.2"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
